Make ptb adapter class for jobstore.
