+++ ptbcontrib/ptb_jobstores/requirements.txt
@@ -1,3 +1,3 @@
 python-telegram-bot[job-queue]~=20
 pymongo>=4.1,<5
-
+SQLAlchemy==1.4.46
