--- ptbcontrib/ptb_mongodb_jobstore/__init__.py
@@ -1,20 +1,0 @@
-# A library containing community-based extension for the python-telegram-bot library
-# Copyright (C) 2020-2023
-# The ptbcontrib developers
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Lesser Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Lesser Public License for more details.
-#
-# You should have received a copy of the GNU Lesser Public License
-# along with this program.  If not, see [http://www.gnu.org/licenses/].
-"""This module contains PTBMongoDBJobStore."""
-from .jobstore import PTBMongoDBJobStore
-
-__all__ = ["PTBMongoDBJobStore"]
